// mobile/src/types.ts
export interface BedConfig {
  id: string
  name: string
  x: number
  z: number
  width: number
  depth: number
  color: string
  sunNeeds: 'full' | 'partial' | 'shade-tolerant'
  plantType?: PlantType
  rotation?: number
  hasTrellis?: boolean
  trellisHeight?: number
}
// ...
export type PlantType =
  | 'tomatoes'
  | 'peppers'
  | 'cucumbers'
  | 'zucchini'
  | 'eggplant'
  | 'squash-winter'
  | 'beans'
  | 'peas'
  | 'bush-beans'
  | 'kale'
  | 'lettuce'
  | 'spinach'
  | 'swiss-chard'
  | 'arugula'
  | 'bok-choy'
  | 'broccoli'
  | 'cauliflower'
  | 'cabbage'
  | 'brussels-sprouts'
  | 'kohlrabi'
  | 'carrots'
  | 'beets'
  | 'radishes'
  | 'onions'
  | 'garlic'
  | 'potatoes'
  | 'herbs'
  | 'basil'
  | 'cilantro'
  | 'parsley'
// ...
export type StructureType = 'building' | 'fence'

export interface Structure {
  id: string
  type: StructureType
  name: string
  position: [number, number, number]
  size: [number, number, number]
  color: string
  castShadow: boolean
  receiveShadow: boolean
}